requirements/app.txt
Django==1.2.1
django-taggit==0.8.0